I... auth is hard. Anyway - local storage solution
I'll definitely need to go through and tidy this up as I tried a few solutions just to get to this state. At any rate, after login the server calls function to save the access token in local storage. More work will need to be done to ensure these are sent when making request to the server via API.